Combine cucumber and cucumber:boot into one job
MSP-11671

The run times for cucumber are so much faster now that `spec --tag
content` is actually the slowest job (on 1.9.3).  The job is so slow
that the two cucumbers can be merged together and they will take the
same time, but we'll consume 2 fewer job slots (1 for 1.9.3 and 1 for
2.1), so the longer per job time is better as long as it does become the
longest job time.