Add a link to PR #4333 in the message
That way, people can see why this is if they care to.

While we have a link to the blog post at http://r-7.co/msfvenom-2015 , I
fear that by the time spring of 2015 rolls around, we'll be on a
different blogging platform and the link may be busted. I'm certain
we'll still be on GitHub then, though.