Fix bug preventing updates in pro
The RPC code previously tested return values to see if an error key was
equal to true.  In fd3b6e7f544373d07078649c267a4d90eb92ba47, this was
changed to check if the error key was truthy.

The pro updater returns an error key with a string describing an error
(or an empty string for no error).  Ruby strings evaluate as truthy
which caused the new behavior to throw an error.  This prevented
checking for and applying updates.

This reverts to the original behavior of checking for true explicitly
instead of checking for a truthy value.

MSP-12235