Address net-dns additional records empty array
Conversion between packet formats can create empty additional
answers fields, which net-dns cannot handle.

Update net-dns' packet parser to be able to deal with empty arrays
such that it doesn't try to call :data on a nil, which is the only
element of an empty array.

Props to mubix for identifying this issues.