Fix #8972, Remove libsodium until packaged better
Due to issues with packaging for multiple platforms libsodium as an optional dependency for 'rbnacl' is being removed.  Once packaging issue are resolved this will be restored.  This removes support for `ed25519` keys used with ssh for the time being however manual installation of this gem allows user to workaround this limitation.